Dev (#3)
* feat: Added nunit tests for running siat probes

* feat: running with codigoPuntoVentaSpecified

* feat: sync working with puntoventa

* test: Stage 2 complete!

* test: Getting cufd

* test: wip Emision individual

* test: Facturas.

* test: Eventos Significativos

* test: eventos significativos

* feat: ci/cd updated src folder

+++ src/Codigos/DocumentoIdentidad.cs
@@ -1,6 +1,6 @@
 namespace Markind.Siat;
 
-public enum DocumentoIdentidad : short
+public enum DocumentoIdentidad : byte
 {
     Ninguno = 0,
     CarnetDeIdentidad = 1,
